Utils.Time: extracted from Utils
