+++ test/hs/Test/Ganeti/Utils.hs
@@ -46,7 +46,6 @@ import Data.List.NonEmpty (NonEmpty((:|)))
 import Data.List
 import Data.Maybe (listToMaybe)
 import qualified Data.Set as S
-import System.Time
 import qualified Text.JSON as J
 #ifdef VERSION_regex_pcre
 import Text.Regex.PCRE
@@ -285,53 +284,6 @@ case_new_uuid_regex = do
   assertBool "newUUID" $ uuid =~ C.uuidRegex
 #endif
 
-prop_clockTimeToString :: Integer -> Integer -> Property
-prop_clockTimeToString ts pico =
-  clockTimeToString (TOD ts pico) ==? show ts
-
-genPicoseconds :: Gen Integer
-genPicoseconds = choose (0, 999999999999)
-
-genTimeDiff :: Gen Utils.TimeDiff
-genTimeDiff = Utils.TimeDiff <$> arbitrary <*> genPicoseconds
-
-genClockTime :: Gen ClockTime
-genClockTime = TOD <$> choose (946681200, 2082754800) <*> genPicoseconds
-
-prop_addToClockTime_identity :: Property
-prop_addToClockTime_identity = forAll genClockTime addToClockTime_identity
-
-addToClockTime_identity :: ClockTime -> Property
-addToClockTime_identity a =
-  Utils.addToClockTime Utils.noTimeDiff a ==? a
-
--- | Verify our work-around for ghc bug #2519. Taking `diffClockTimes` form
--- `System.Time`, this test fails with an exception.
-prop_timediffAdd :: Property
-prop_timediffAdd =
-  forAll genClockTime $ \a ->
-  forAll genClockTime $ \b ->
-  forAll genClockTime $ \c -> timediffAdd a b c
-
-timediffAdd :: ClockTime -> ClockTime -> ClockTime -> Property
-timediffAdd a b c =
-  let fwd = Utils.diffClockTimes a b
-      back = Utils.diffClockTimes b a
-  in Utils.addToClockTime fwd (Utils.addToClockTime back c) ==? c
-
-prop_timediffAddCommutative :: Property
-prop_timediffAddCommutative =
-  forAll genTimeDiff $ \a ->
-  forAll genTimeDiff $ \b ->
-  forAll genClockTime $ \c -> timediffAddCommutative a b c
-
-timediffAddCommutative ::
-  Utils.TimeDiff -> Utils.TimeDiff -> ClockTime -> Property
-timediffAddCommutative a b c =
-  Utils.addToClockTime a (Utils.addToClockTime b c)
-  ==?
-  Utils.addToClockTime b (Utils.addToClockTime a c)
-
 
 -- | Test normal operation for 'chompPrefix'.
 --
@@ -430,10 +382,6 @@ testSuite "Utils"
 #ifdef VERSION_regex_pcre
             , 'case_new_uuid_regex
 #endif
-            , 'prop_clockTimeToString
-            , 'prop_addToClockTime_identity
-            , 'prop_timediffAdd
-            , 'prop_timediffAddCommutative
             , 'prop_chompPrefix_normal
             , 'prop_chompPrefix_last
             , 'prop_chompPrefix_empty_string
